EBNT-321 Decoupled tests changed / added host_only param to delete images only from image host

+++ src/ebonite/client/base.py
@@ -138,7 +138,7 @@ class Ebonite:
         image.model = model
         return self.meta_repo.create_image(image)
 
-    def delete_image(self, image: Image, environment: RuntimeEnvironment = None):
+    def delete_image(self, image: Image, environment: RuntimeEnvironment = None, host_only: bool = False):
         """
         Deletes existing image from metadata repository and image provider
 
@@ -149,7 +149,9 @@ class Ebonite:
             environment = self.get_default_environment()
 
         environment.params.remove_image(image)
-        self.meta_repo.delete_image(image)
+        if not host_only:
+            self.meta_repo.delete_image(image)
+        return True
 
     def get_image(self, name: str, model: Model) -> Image:
         """
